feat(pages): reorganize imports and fix module references
Reorganized imports in various files within the pages module to improve readability and maintain consistency. Fixed module references in multiple files to ensure proper functionality and maintainability.
