+++ freestream/pages/utils/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from .utility_funcs import *
+from .styles import *
